Don't embed structs (apparently it changes the HJSON structure)
