--- src/config/admin.go
@@ -1,5 +1,0 @@
-package config
-
-type AdminConfig struct {
-	AdminListen string `comment:"Listen address for admin connections. Default is to listen for local\nconnections either on TCP/9001 or a UNIX socket depending on your\nplatform. Use this value for yggdrasilctl -endpoint=X. To disable\nthe admin socket, use the value \"none\" instead."`
-}
